Migrate the shadow modifiers to dc_bundle
