+++ crates/dc_bundle/src/legacy_definition/modifier.rs
@@ -16,5 +16,6 @@
 
 pub mod blend;
 pub mod filter;
+pub mod shadow;
 pub mod text;
 pub mod transform;
